move promts to file for cloudfoundry subgen
